# main/openapi.yaml
openapi: "3.0.3"
info:
  title: SylLab API
  description: |
    Anti-Plagiarism Code Forensics Platform API.
    Establishes proctored authentic baselines and detects AI-assisted code generation
    through sophistication delta analysis, technique genealogy validation, and cohort comparison.
  version: "1.0.0"
  contact:
    name: SylLab Support
    email: support@syllab.io
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.syllab.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
            example:
              email: "student@university.edu"
              password: "SecurePass123!"
              firstName: "Amir"
              lastName: "Seitkali"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "instructor@university.edu"
              password: "SecurePass123!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expiresIn: 900
                user:
                  id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  email: "instructor@university.edu"
                  role: "INSTRUCTOR"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate refresh token
      responses:
        "204":
          description: Logged out successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # USERS
  # ============================================================
...
  /courses:
    get:
      tags: [Courses]
      summary: List courses accessible to the current user
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortParam"
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCourses"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Courses]
      summary: Create a new course (Instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
            example:
              name: "Backend Development 2025"
              code: "CS301-2025"
              weeklyTargets:
                "1": 20
                "2": 25
                "3": 30
                "4": 35
                "5": 40
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /courses/{courseId}/enroll:
    post:
      tags: [Courses]
      summary: Enroll a student in a course (Instructor/Admin)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId]
              properties:
                studentId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # SESSIONS (Proctored)
  # ============================================================
...
  /baselines:
    post:
      tags: [Baselines]
      summary: Submit Week 1 baseline code (Student only, during active session)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sessionCode, file]
              properties:
                sessionCode:
                  type: string
                  example: "ABC123"
                file:
                  type: string
                  format: binary
                  description: JavaScript or TypeScript source file (max 5MB)
      responses:
        "201":
          description: Baseline submitted and analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /baselines/{studentId}/{courseId}:
    get:
      tags: [Baselines]
      summary: Get baseline for a student in a course (Instructor/Admin or own baseline)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Baseline details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # ASSIGNMENTS
  # ============================================================
  /assignments:
    get:
      tags: [Assignments]
      summary: List assignments for a course
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortParam"
        - name: courseId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: weekNumber
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 16
      responses:
        "200":
          description: Paginated assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAssignments"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Assignments]
      summary: Create a new assignment (Instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
            example:
              courseId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              title: "Lab 4 - Error Handling & Service Layer"
              weekNumber: 4
              expectedScore: 45.0
              dueDate: "2025-10-05T23:59:00Z"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # SUBMISSIONS
  # ============================================================
  /submissions:
    get:
      tags: [Submissions]
      summary: List submissions for an assignment (Instructor) or own submissions (Student)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortParam"
        - name: assignmentId
          in: query
          schema:
            type: string
            format: uuid
        - name: studentId
          in: query
          schema:
            type: string
            format: uuid
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [NORMAL, MONITOR, FLAGGED, CRITICAL]
      responses:
        "200":
          description: Paginated submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubmissions"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Submissions]
      summary: Submit code for an assignment (Student only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [assignmentId, file]
              properties:
                assignmentId:
                  type: string
                  format: uuid
                file:
                  type: string
                  format: binary
                  description: JavaScript or TypeScript source file (max 5MB)
      responses:
        "201":
          description: Submission accepted and analysis triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # ANALYSIS
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Obtain via POST /auth/login. Include as: Authorization: Bearer <token>"
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for offset pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: "-createdAt"
      description: "Sort field. Prefix with - for descending. Example: -createdAt, sophisticationScore"
  responses:
    BadRequest:
      description: Bad request — invalid parameters or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 400
            error: "Bad Request"
            message: "sessionCode must be a 6-character alphanumeric string"
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            error: "Unauthorized"
            message: "Invalid or expired access token"
    Forbidden:
      description: Forbidden — insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 403
            error: "Forbidden"
            message: "Only INSTRUCTOR role can perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 404
            error: "Not Found"
            message: "Submission with id 'abc-123' not found"
    Conflict:
      description: Conflict — resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 409
            error: "Conflict"
            message: "Student already has a baseline for this course"
    UnprocessableEntity:
      description: Unprocessable Entity — validation errors
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            statusCode: 422
            error: "Unprocessable Entity"
            message: "Validation failed"
            details:
              - field: "weekNumber"
                message: "weekNumber must be between 1 and 16"
              - field: "expectedScore"
                message: "expectedScore must be a number between 0 and 100"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred. Please try again later."
  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 87
        totalPages:
          type: integer
          example: 5
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserResponse"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        instructorId:
          type: string
          format: uuid
        weeklyTargets:
          type: object
          additionalProperties:
            type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateCourseRequest:
      type: object
      required: [name, code, weeklyTargets]
      properties:
        name:
          type: string
          maxLength: 200
        code:
          type: string
          pattern: "^[A-Z]{2,6}[0-9]{3}-[0-9]{4}$"
        weeklyTargets:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 100
    PaginatedCourses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        enrolledAt:
          type: string
          format: date-time
...
    AstMetrics:
      type: object
      description: 15+ AST-derived sophistication metrics
      properties:
        errorHandlingTier:
          type: integer
          minimum: 0
          maximum: 5
          description: "0=none, 1=basic try-catch, 3=specific exceptions, 5=custom hierarchies"
        architectureTier:
          type: integer
          minimum: 0
          maximum: 5
          description: "0=flat functions, 2=simple classes, 4=service layer, 5=repository/CQRS"
        typeSafetyTier:
          type: integer
          minimum: 0
          maximum: 3
          description: "0=none, 1=partial hints, 3=comprehensive TypeScript"
        hasDecorators:
          type: boolean
        hasAsyncAwait:
          type: boolean
        hasContextManagers:
          type: boolean
        hasMetaclasses:
          type: boolean
        hasCircuitBreaker:
          type: boolean
        hasDependencyInjection:
          type: boolean
        namingVerbosityScore:
          type: number
          minimum: 0
          maximum: 2
        functionCount:
          type: integer
        classCount:
          type: integer
        averageFunctionLength:
          type: number
        nestingDepth:
          type: integer
        commentRatio:
          type: number
    BaselineResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        contentHash:
          type: string
          description: SHA-256 hash of submitted code
        sophisticationScore:
          type: number
          minimum: 0
          maximum: 100
        metrics:
          $ref: "#/components/schemas/AstMetrics"
        isLocked:
          type: boolean
        submittedAt:
          type: string
          format: date-time
    AssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        title:
          type: string
        weekNumber:
          type: integer
        expectedScore:
          type: number
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateAssignmentRequest:
      type: object
      required: [courseId, title, weekNumber, expectedScore, dueDate]
      properties:
        courseId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 300
        weekNumber:
          type: integer
          minimum: 1
          maximum: 16
        expectedScore:
          type: number
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date-time
    PaginatedAssignments:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    SubmissionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assignmentId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        contentHash:
          type: string
        sophisticationScore:
          type: number
        metrics:
          $ref: "#/components/schemas/AstMetrics"
        submittedAt:
          type: string
          format: date-time
        analysisStatus:
          type: string
          enum: [PENDING, COMPLETE, FAILED]
    PaginatedSubmissions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...